Move monkey patches to arabot.core `Message.rsearch` is now `Context.rsearch`
