+++ arabot/cogs/google/tts.py
@@ -40,8 +40,7 @@ class TextToSpeech(Cog, category=Category.GENERAL, keys={"g_tts_key"}):
         langs = await self.voices()
         lang, text = self.parse_query(query, langs)
 
-        ctx.message.content = ""
-        if not text and not (text := await ctx.message.rsearch(ctx, "content")):
+        if not text and not (text := await ctx.rsearch("content")):
             await ctx.send("I need text to pronounce")
             return
